[plugins] Use 64-bit types for some of the key plugin fields
This should match the Flutter GTK embedder implementation (as of
v3.22.0).

+++ flutter-plugins/src/keyevent.rs
@@ -48,9 +48,9 @@ pub struct KeyAction {
     pub scan_code: i32,
     pub modifiers: i32,
     #[serde(rename = "specifiedLogicalKey")]
-    pub specified_logical_key: i32,
+    pub specified_logical_key: i64,
     #[serde(rename = "unicodeScalarValues")]
-    pub unicode_scalar_values: i32,
+    pub unicode_scalar_values: i64,
     pub keymap: String,
     #[serde(rename = "type")]
     pub _type: KeyActionType,
